+++ server/schemas/typeDefs.js
@@ -29,7 +29,7 @@ type Query {
   type Mutation {
     addUser(username: String!, email: String!, password: String!): Auth
     login(email: String!, password: String!): Auth
-    addBet(betType: String!, matchup: String!, spread: Int, winner: String, total: Int, endTime: Int!, betStatus: String!, units: Int!): Bet
+    addBet(betType: String!, matchup: String!, spread: Float, winner: String, total: Float, endTime: Int!, betStatus: String!, units: Int!): Bet
 
   }
 `;
